+++ src/common/index.ts
@@ -3,4 +3,5 @@ export * from "./json-file";
 export * from "./lodash";
 export * from "./logger";
 export * from "./pubkey";
+export * from "./utility";
 export * from "./web3";