+++ src/entity/bn.ts
@@ -1,0 +1,14 @@
+import BN from "bn.js";
+
+// round up
+export function divCeil(a: BN, b: BN): BN {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  const dm = a.divmod(b);
+
+  // Fast case - exact division
+  if (dm.mod.isZero()) return dm.div;
+
+  // Round up
+  return dm.div.negative !== 0 ? dm.div.isubn(1) : dm.div.iaddn(1);
+}