+++ src/liquidity/type.ts
@@ -1,7 +1,8 @@
 import { JsonFileMetaData } from "../common";
 import { LpTokenInfo } from "../token";
 
-export type LiquidityVersion = 2 | 3 | 4;
+// 2 | 3 | 4
+export type LiquidityVersion = 4;
 
 export interface LiquidityPoolBaseInfo {
   readonly id: string;