+++ src/serum/type.ts
@@ -1,1 +1,2 @@
-export type SerumVersion = 1 | 2 | 3;
+// 1 | 2 | 3
+export type SerumVersion = 3;