+++ src/swap/swap.ts
@@ -1,3 +1,0 @@
-export class Swap {
-  static getAmountOut() {}
-}